requirements.txt
Django==1.11
django-wysiwyg-redactor==0.5.1
Fabric3==1.13.1.post1
Jinja2==2.9.6
psycopg2==2.7.1